refactor: Fix folder structure of ai-brainz
